+++ libs/angular-components/tedi/components/buttons/collapse/collapse.component.ts
@@ -5,7 +5,10 @@ import {
   AfterViewInit,
   ChangeDetectionStrategy,
   ViewEncapsulation,
+  inject,
+  Renderer2
 } from "@angular/core";
+import { TediTranslationPipe } from "../../../services/translation/translation.pipe";
 import { IconComponent } from "../../base/icon/icon.component";
 import { TextComponent } from "../../base/text/text.component";
 
@@ -14,24 +17,22 @@ export type ArrowType = "default" | "secondary";
 @Component({
   standalone: true,
   selector: "tedi-collapse",
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  encapsulation: ViewEncapsulation.None,
-  imports: [IconComponent, TextComponent],
+  imports: [IconComponent, TextComponent, TediTranslationPipe],
   templateUrl: "./collapse.component.html",
   styleUrls: ["./collapse.component.scss"],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  encapsulation: ViewEncapsulation.None
 })
 export class CollapseComponent implements AfterViewInit {
   /**
    * The title/header element for the collapsible section.
    * Rendered inside the toggle button.
-   * @default "Näita"
    */
-  openText = input<string>("Näita");
+  openText = input<string>();
   /**
    * Text shown on the toggle button when the content is expanded.
-   * @default "Peida"
    */
-  closeText = input<string>("Peida");
+  closeText = input<string>();
   /**
    * Whether the collapse should be initially open.
    * @default false
@@ -48,6 +49,7 @@ export class CollapseComponent implements AfterViewInit {
    */
   arrowType = input<ArrowType>("default");
 
+  renderer = inject(Renderer2);
   collapseContentId: string = `collapse-content-${self.crypto.randomUUID()}`;
   isOpen = signal<boolean>(false);